requirements.txt
anthropic>=0.8.0
python-dotenv>=1.0.0
yfinance>=0.2.35
newsapi-python>=0.2.7
alpaca-py>=0.10.0
mcp
slack_sdk
mcp-python-sdk>=0.1.0
slack-sdk>=3.19.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
black>=22.0.0
isort>=5.10.0
build>=0.7.0
twine>=4.0.0